feat(header): extract rotating job header into component
Move the rotating job-title header out of pages/index into a new
components/Header.js and wire it into the home page. Header now
manages its own job title state and interval with useEffect, receiving
an optional jobTitles prop (defaulting to the previous list). This
removes duplicated state from pages/index and simplifies the page
component.

Also clean up minor formatting and error handling:
- Replace inline header markup in pages/index with <Header /> import.
- Remove the old jobTitles constant and interval logic from pages/index.
- Normalize some object/JSX formatting and improve error message
  handling in the API submission logic.

These changes improve component separation, readability, and reuse.
