+++ src/pages/index.js
@@ -1,31 +1,16 @@
 import Head from "next/head";
-import { useState, useEffect } from "react";
+import { useState } from "react";
 
 import { Inter } from "next/font/google";
 
 import Footer from "components/Footer";
+import Header from "components/Header";
 import ApiResponse from "components/apiResponse";
 
 import { BeakerIcon } from "@heroicons/react/24/solid";
 import { usePlausible } from "next-plausible";
 
 const inter = Inter({ subsets: ["latin"] });
-const jobTitles = [
-  "Nurse Practioner",
-  "Software engineer",
-  "Aircraft Mechanic",
-  "Cosmetologist",
-  "Epic Analyst",
-  "Zookeeper",
-  "Cashier",
-  "Dental Hygienist",
-  "Dental Assistant",
-  "Medical Assistant",
-  "Pharmacy Technician",
-  "Physical Therapist",
-  "Radiologic Technologist",
-  "Registered Nurse",
-];
 
 function oldformatResponse(response) {
   const parts = response.split("\n\n");
@@ -55,20 +40,11 @@ function oldformatResponse(response) {
 
 export default function Home() {
   const plausible = usePlausible();
-  const [jobTitleIndex, setJobTitleIndex] = useState(0);
   const [jobName, setjobName] = useState("");
   const [result, setResult] = useState();
   const [apiResponseContent, setApiResponseContent] = useState("");
   const [errorMessage, setErrorMessage] = useState("");
   const [isProcessing, setIsProcessing] = useState(false);
-
-  useEffect(() => {
-    const intervalId = setInterval(() => {
-      setJobTitleIndex((jobTitleIndex + 1) % jobTitles.length);
-    }, 1000);
-
-    return () => clearInterval(intervalId);
-  }, [jobTitleIndex]);
 
   async function onSubmit(event) {
     event.preventDefault();
@@ -88,12 +64,17 @@ export default function Home() {
 
       const data = await response.json();
       if (response.status !== 200) {
-        throw data.error || new Error(`Request failed with status ${response.status}`);
+        throw (
+          data.error ||
+          new Error(`Request failed with status ${response.status}`)
+        );
       }
       setApiResponseContent(data.result);
     } catch (error) {
       console.error(error);
-      setErrorMessage(error.message || "Something went wrong. Please try again.");
+      setErrorMessage(
+        error.message || "Something went wrong. Please try again.",
+      );
     } finally {
       setIsProcessing(false);
     }
@@ -102,25 +83,16 @@ export default function Home() {
     <>
       <Head>
         <title>Job Descriptions</title>
-        <meta name="description" content="Find out what someone with a job title does.." />
+        <meta
+          name="description"
+          content="Find out what someone with a job title does.."
+        />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
 
         <link rel="icon" href="/favicon.ico" />
       </Head>
       <div className="min-h-screen bg-gray-100">
-        <header className="bg-white py-6 h-50">
-          <h1 className="text-2xl text-center font-bold">
-            Ever wonder what a <br />
-            <span className="inline-block">
-              {jobTitles.map((title, index) => (
-                <span key={title} className={index === jobTitleIndex ? "" : "hidden"}>
-                  {title}
-                </span>
-              ))}
-            </span>{" "}
-            does?
-          </h1>
-        </header>
+        <Header />
 
         <main className="flex flex-col items-center pt-12">
           <div className="w-full sm:w-2/3 md:w-1/2 flex flex-col items-center">
